requirements.txt
Flask==2.0.1
requests==2.25.1
python-dateutil==2.8.1
gradio==2.3.9
ultrabot==0.1.0
python-dotenv==1.0.0
